requirements.txt
gitpython
pygithub
python-gitlab
